ai-service/requirements.txt
Flask==3.0.3
flask-limiter==3.8.0
python-dotenv==1.0.1
requests==2.32.3
Limits==3.13.0
